2021 02 04 BoradMapperTests, FreeBoardMapper.java.xml FreeBoardVO 코드수정/ lombok-build path 라이브러리 추가

+++ src/main/java/org/zerock/domain/FreeBoardVO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class FreeBoardVO {
-	private int no;
+	private Long no;
 	private String title;
 	private String content;
 	private Date regdate;
